Include admin view
The admin view allows a superuser to get an overview of every existing
dashboard, showing metrics like the total number of tasks, the completed
ones, etc.
Also, shows a general status of the server and, in next version, stats
about the workers.

In addition, the model CompletedTask now includes a field to indicate if
it is the last completed task.

Related to: https://gitlab.com/cauldron2/cauldron-feedback/issues/38
Related to: https://gitlab.com/cauldron2/mordred/merge_requests/14