[hatstall] Hatstall is now an application that can be optionally included in Cauldron
These changes allows to deploy Hatstall in Cauldron as an application in the same
container.
Hatstall will use the same authentication as Cauldron. Only admin users will be able to
edit the affiliations.
For the authentication system, a new middleware has been created named HatstallAuthorizationMiddleware
that is only active when Cauldron is deployed with Hatstall.