Set the x-axis ranges to those filtered by the date picker.
This commit updates the way the parameters are passed to the functions.
Now it receives a datetime, and it is transformed to str for Elasticsearch.

It also set the range for the Bokeh visualizations to the ranged passed to the function.

Finally replaces the y-axis to not be scientific notation when the values are very high